splitting annotation template PR in two - base panel annotator
